Fix class not found paper conflict
