+++ src/main/java/ch/njol/skript/bukkitutil/EntityUtils.java
@@ -18,34 +18,24 @@
  */
 package ch.njol.skript.bukkitutil;
 
-import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.entity.Ageable;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
-import org.bukkit.entity.LivingEntity;
-import org.bukkit.entity.Mob;
 import org.bukkit.entity.Piglin;
-import org.bukkit.entity.Player;
 import org.bukkit.entity.Zoglin;
 import org.bukkit.entity.Zombie;
-import org.bukkit.util.Vector;
-import org.eclipse.jdt.annotation.Nullable;
 
 import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
 
 import ch.njol.skript.Skript;
 import ch.njol.skript.entity.EntityData;
-import io.papermc.paper.entity.LookAnchor;
 
 /**
  * Utility class for quick {@link Entity} methods
  */
 public class EntityUtils {
-
-	public static final boolean LOOK_ANCHORS = Skript.classExists("io.papermc.paper.entity.LookAnchor");
-	public static final boolean LOOK_AT = Skript.methodExists(Mob.class, "lookAt", Entity.class);
 
 	private static final boolean HAS_PIGLINS = Skript.classExists("org.bukkit.entity.Piglin");
 
@@ -59,83 +49,6 @@ public class EntityUtils {
 			Class<? extends Entity> c = e.getEntityClass();
 			if (c != null)
 				SPAWNER_TYPES.put(EntityData.fromClass(c), e);
-		}
-	}
-
-	/**
-	 * Instruct a Mob (1.17+) to look at a specific vector/location/entity.
-	 * Object can be a {@link org.bukkit.util.Vector}, {@link org.bukkit.Location} or {@link org.bukkit.entity.Entity}
-	 * 
-	 * @param target The vector/location/entity to make the livingentity look at.
-	 * @param entities The living entities to make look at something.
-	 */
-	public static void lookAt(Object target, LivingEntity... entities) {
-		lookAt(target, null, null, entities);
-	}
-
-	/**
-	 * Instruct a Mob (1.17+) to look at a specific vector/location/entity.
-	 * Object can be a {@link org.bukkit.util.Vector}, {@link org.bukkit.Location} or {@link org.bukkit.entity.Entity}
-	 * 
-	 * @param target The vector/location/entity to make the livingentity look at.
-	 * @param headRotationSpeed The rotation speed at which the living entities will rotate their head to the target. Vanilla default values range from 10-50. Doesn't apply to players.
-	 * @param maxHeadPitch The maximum pitch at which the eyes/feet can go to. Doesn't apply to players.
-	 * @param entities The living entities to make look at something.
-	 */
-	public static void lookAt(Object target, @Nullable Float headRotationSpeed, @Nullable Float maxHeadPitch, LivingEntity... entities) {
-		if (target == null || !LOOK_AT)
-			return;
-		// Use support for players if using Paper 1.19.1+
-		if (LOOK_ANCHORS) {
-			lookAt(LookAnchor.EYES, headRotationSpeed, maxHeadPitch, entities);
-			return;
-		}
-		for (LivingEntity entity : entities) {
-			if (!(entity instanceof Mob))
-				continue;
-			Mob mob = (Mob) entity;
-			Bukkit.getMobGoals().getRunningGoals(mob, com.destroystokyo.paper.entity.ai.GoalType.LOOK).forEach(goal -> Bukkit.getMobGoals().removeGoal(mob, goal));
-			float speed = headRotationSpeed != null ? headRotationSpeed : mob.getHeadRotationSpeed();
-			float maxPitch = maxHeadPitch != null ? maxHeadPitch : mob.getMaxHeadPitch();
-			Bukkit.getMobGoals().addGoal(mob, 0, new LookGoal(target, mob, speed, maxPitch));
-		}
-	}
-
-	/**
-	 * Instruct a Mob (1.17+) or Players (1.19.1+) to look at a specific vector/location/entity.
-	 * Object can be a {@link org.bukkit.util.Vector}, {@link org.bukkit.Location} or {@link org.bukkit.entity.Entity}
-	 * THIS METHOD IS FOR 1.19.1+ ONLY. Use {@link} otherwise.
-	 * 
-	 * @param entityAnchor What part of the entity the player should face assuming the LivingEntity argument contains a player. Only for players.
-	 * @param target The vector/location/entity to make the livingentity or player look at.
-	 * @param headRotationSpeed The rotation speed at which the living entities will rotate their head to the target. Vanilla default values range from 10-50. Doesn't apply to players.
-	 * @param maxHeadPitch The maximum pitch at which the eyes/feet can go to. Doesn't apply to players.
-	 * @param entities The living entities to make look at something. Players can be involved in 1.19.1+
-	 */
-	public static void lookAt(LookAnchor entityAnchor, Object target, @Nullable Float headRotationSpeed, @Nullable Float maxHeadPitch, LivingEntity... entities) {
-		if (target == null || !LOOK_AT || !LOOK_ANCHORS)
-			return;
-		for (LivingEntity entity : entities) {
-			if (entity instanceof Player) {
-				Player player = (Player) entity;
-				if (target instanceof Vector) {
-					Vector vector = ((Vector)target);
-					player.lookAt(vector.getX(), vector.getY(), vector.getZ(), LookAnchor.EYES);
-					player.lookAt(vector.getX(), vector.getY(), vector.getZ(), LookAnchor.FEET);
-				} else if (target instanceof Location) {
-					player.lookAt((Location) target, LookAnchor.EYES);
-					player.lookAt((Location) target, LookAnchor.FEET);
-				} else if (target instanceof Entity) {
-					player.lookAt((Entity) target, LookAnchor.EYES, entityAnchor);
-					player.lookAt((Entity) target, LookAnchor.FEET, entityAnchor);
-				}
-			} else if (entity instanceof Mob) {
-				Mob mob = (Mob) entity;
-				Bukkit.getMobGoals().getRunningGoals(mob, com.destroystokyo.paper.entity.ai.GoalType.LOOK).forEach(goal -> Bukkit.getMobGoals().removeGoal(mob, goal));
-				float speed = headRotationSpeed != null ? headRotationSpeed : mob.getHeadRotationSpeed();
-				float maxPitch = maxHeadPitch != null ? maxHeadPitch : mob.getMaxHeadPitch();
-				Bukkit.getMobGoals().addGoal(mob, 0, new LookGoal(target, mob, speed, maxPitch));
-			}
 		}
 	}
